It adds empty lines before script. Refactoring. Code cleanup. Comments were added.

+++ script.h
@@ -24,8 +24,8 @@ typedef struct
     TTF_Font        *ttf_font;
     SDL_Color       sdl_text_color;
     linkedList_t    wrappedScriptList;              /* Linked list of wrapped lines */
-    uint32_t        wrappedScriptHeightPx;          /* Height of one line */
-    uint32_t        heightOffsetPx;                 /* Offset inside on line. Range: 0 .. wrappedScriptHeightPx - 1 */
+    uint16_t        wrappedScriptHeightPx;          /* Height of one line */
+    uint16_t        heightOffsetPx;                 /* Offset inside on line. Range: 0 .. wrappedScriptHeightPx - 1 */
 } wrappedScript_t;
 
 #endif /* INCLUDE_SCRIPT_H */
